Fix #528, invite > accept > leave > invite > accept flow being broken
The problem was that in order to refresh invitations it would search for
invitations to the same voice/organization and just use the
invitation_request_id if it already existed previously.  But it didn't
check if it was already a close invitation, thus leaving the
InvitationRequest unexistent.

Now it checks to make sure it's not invitation_{accepted,rejected}*
before trying to use the invitation_request_id from it.