Improve search results for global search
Before these changes to find "cersei" you had to search for "cersei",
"cer" wouldn't give you any results.  With these new changes "cer" will
give you Cersei related results.

---

In order to make it support I turned the plainto_tsquery into a normal
to_tsquery, but I formatted the input I give it specially with:

query.split(/[ ]+/).join(':* | ') + ':*'

The | could be changed into & to change to change from an OR behaviour
to an AND behaviour.

---

This approach still uses vectors, thus is only compatible with ASCII
stuff, as far as I can tell.  This is quick but does not support search
for UTF-8 languages, like Arab, Japanese etc.

For now this'll have to do, in the future I need to look into supporting
all languages, maybe with LIKE or ILIKE but that has performance
implications, so yeah need to look into it.