Fix 656, requirements when publishing voice
If voice is STATUS_PUBLISHED or STATUS_UNLISTED then it has to have a
background image and at least 15 approved posts, otherwise it's a 403.

Also on the VoicesController.update fixed an issue where it wasn't
checking if there was a Twitter search term before running the Twitter
fetcher, thus causing errors.