Fix ticket 697
Fixes the NotificationMailer's handling of the receiver entity being an
Anonymous entity.

The error was about it trying to find NotificationSettings for an
anonymous entity, thus creating an error when it tried to use the value
it couldn't find.

Now if the receiver is an anonymous entity then it'll use the settings
of the real person/entity, handling is still the same if it's an
organization.

Signed-off-by: Sergio de la Garza <sergio@delagarza.io>