Use Mozilla's Readability lib
This is in order to not depend on a service and instead use an open
source version whose output we can control and manipulate freely.

Also avoid any quota limits or required payments in the future.

This patch adds a ReadabilityParser class, one instance manages one URL
that can be fetched and processed with the methods provided by the class.

Signed-off-by: Sergio de la Garza <sergio@delagarza.io>