Fix 690, Voice update fields not emptyable
This fixes an issue with Multer (multipart middleware) where it was
remvoing empty (i.e. length 0) fields from the request's body, thus not
alowing the back end to see that this field wanted to be updated to an
empty value.

This also fixes an issue with default management in the Voice's update
endpoint where it just feel back if the req.body's value was falsey, and
an empty string is falsey, so it just got ignored.

Signed-off-by: Sergio de la Garza <sergio@delagarza.io>