fix / create new posts show no initial datetime
since the last change made to the EditablePost widget, it expect that
the model data passed contains a `publishedAt` property, new posts does
not contain that property so the rome external lib is not handling the
initial setup correctly.

the problem with it is that we are not showing the datetime on the ui.