Add tests for K.Voice
And update the bin/posts_generator.js script for the posts to have
random URLs and for the approved attribute to be random (true,false).
The Votes and SavedPosts also now only apply to the posts of the Voice
you provide, not every post.

Updated one of the K.Post hook functions which was trying to call 'res'
(resolve) but 'res' was being over-written as 'res' (response).

In addition optimized a K.Voice method.